% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCube.R
\name{as.data.frame.HGSArray}
\alias{as.data.frame.HGSArray}
\title{Convert HGSArray to a data.frame}
\usage{
\method{as.data.frame}{HGSArray}(x)
}
\arguments{
\item{x}{An HGSArray S3 object.}
}
\value{
A data.frame where columns represent variables from the HGSArray.
  Column names are variable names.  A "NodeID" or "CellID" variable in the
  HGSArray, if present, is not included in the data.frame, but values are
  used as row names the resulting data.frame.
}
\description{
Create a data.frame containing values from an HGSArray.
}
