% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGSFile.R
\name{HGSFile}
\alias{HGSFile}
\title{Read HGS output file and create HSGFile object}
\usage{
HGSFile(x = file.choose())
}
\arguments{
\item{x}{A file name to read.}
}
\value{
Returns an S3 object with a class of "HGSFile".  The object will also
  have a subclass, which is the text "HGSFile_" contatinated with the type of
  HGS output file.  So, for instance, the S3 class resulting from reading an
  HGS "pm" file will be c("HGSFile_pm", "HGSFile").  (To see the currently
  supported types, use \code{\link{HGSSupported}}

  An HGSFile object is a list which has, at minimum, three elements:

  \itemize{

  \item{ \code{fileInfo} will contain the results from
  \code{\link{file.info}} for the HGS file that was read, at the time the
  HGSFile object was created.}

  \item{\code{description} will contain the model run description provided by
  the HGS user when the model was run.}

  \item{\code{taggedLines} is a data.frame with one record for each "tagged"
  line of the HGS file.  Tagged lines demarcate the beginning of an item of
  interest in the HGS data file.  For instance, "pm" files contain tags for
  titles, variable names, zones, data, and node IDs. To see the tags for each
  supported for each file type, use \code{\link{HGSTags}}.}

  }

  An HGSFile object may have additional elements, determined by the subclass.

  \code{HGSFile_pm} objects will contain the following additional elements:

  \itemize{

  \item{\code{variables} is a vector of variable names included in output.}

  \item{\code{elementNodeLookup} A named numeric vector telling the number of
  lines to skip and the number of lines to read in the original data file to
  determine the node IDs associated with each element of the model.  Useful
  with \code{\link{scan}}.  (HGS elements are, e.g., the cubes demarcated by
  eight nodes in the simulation space.)}

  \item{\code{modelDim} A named numeric vector telling the number of nodes in the
  simulation space in X, Y, and Z dimensions.}

  \item{\code{blocks} A list containing sublists describing information about
  each "block" in the original data file. The information available for each
  \code{block} can be investigated with \code{\link{HGSQueryBlocks}}.  The
  information stored in a \code{block}'s sublist is variable, depending on
  the contents of the HGS data file.  For instance, in a 'HGSFile_pm' object,
  there is a block for each simulation time at which the user requested
  output from HGS. Therefore, in a 'HGSFile_pm' object, each block has a
  "SOLUTIONTIME" element.  Critically, each block has a \code{DATAMAP}
  element, containing a data.frame with one row for each variable that can be
  extracted from the HGS data file.  Model variable names are the row names
  of the \code{DATAMAP} data.frame. The columns tell how many lines to skip
  and how many lines to read in the datafile in order to retrieve the data
  for each variable from the \code{block}.  That said, rather than try to
  read the \code{DATAMAP}s directly, use the \code{\link{HGSGetData}}
  function, which uses the \code{DATAMAP}s correctly to retrieve data for
  specific variables and simulation times from an HGS output file.}

  }
}
\description{
Returns information about an file created by HGS in the form of an HGSFile
object.  The HGSFile object can then be passed to other functions to return
HGS output in R friendly formats.
}
\examples{
  # requires name of a legetimate pm file
  HGSF <- HGSFile("a_File_Name.pm.dat")

}
\seealso{
\code{\link{HGSQueryBlocks}}
}
