% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{HGSQueryBlocks}
\alias{HGSQueryBlocks}
\title{Display data associated with output blocks in an HGSFile.}
\usage{
HGSQueryBlocks(HGSFile, descriptors = NULL)
}
\arguments{
\item{HGSFile}{An S3 object of class "HGSFile" created by calling
\code{\link{HGSFile}}.}

\item{descriptors}{Either NULL, or a character vector of descriptor names.  See
documentation of return value, below.}
}
\value{
When \code{descriptors} is \code{NULL}, return value is a \code{list} of
descriptor names available for each block. When \code{descriptors} is a single
descriptor name, return value will be a list of values, one value for each
block.  When \code{descriptors} is a vector of more than one descriptor name,
the return value is a \code{list} of \code{lists} of descriptor values.
}
\description{
When HGS is run, the user requests output at specific solution times for key
data from the model.  For each requested time, there is a block of data in
the output file.  This command displays critical descriptors for each block
of data.
}
\examples{
HGSF <- HGSFile("a_File_Name.pm.dat") # requires name of a legetimate pm file

# returns the names of the information in each query block
HGSQueryBlocks(HGSF)

# returns the simulation time for each block
HGSQueryBlocks(HGSF, "SOLUTIONTIME")

# returns the simulation times for each simulation block followed by the variable locations for each simulation block
HGSQueryBlocks(HGSF, c("SOLUTIONTIME", "VARLOCATION"))

}
