% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{HGSSupported}
\alias{HGSSupported}
\alias{HGSTags}
\title{See supported file types and associated tags.}
\usage{
HGSSupported(quiet = F)

HGSTags(quiet = F)
}
\arguments{
\item{quiet}{When set TRUE, will suppress printing of information to the
screen.}
}
\value{
\code{HGSSupported} returns a character vector of file extensions
  (used by HGS to indicate the type of file).  If a file extension is in this
  vector, the associated HGS file type can be read by this package.

  \code{HGSTags} returns a named list of character vectors where the names
  are the supported file types and the associated vectors contain the tags
  that are processed for each file type, and the data type associated with each tag.

  Values are returned invisibly (without printing, see
  \code{\link{invisible}}).
}
\description{
These functions provide some basic information about the functionality of the HGSReader package.
}
\details{
HGS outputs several different file types, differentiated by the extensions of
the file names.  Each file type has a set of "tags" that occur in the file to
demarcate the beginning of sections of the file that contain different types
of information.  These functions show which file types are supported and which
tags can be processed in each file type.
}
\examples{
  # See supported file types
  HGSSupported()

  # See tags that are processed in each file type
  HGSTags()

}
