% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCube.R
\name{HGSGetData}
\alias{HGSGetData}
\alias{HGSCube}
\title{Get data values from HGS files}
\usage{
HGSGetData(HGSFile, variables, blockNumber = NULL, solutionTime = NULL,
  includeCoords = F)

HGSCube(HGSdf)
}
\arguments{
\item{HGSFile}{An S3 object of class "HGSFile" created by calling
\code{\link{HGSFile}}.}

\item{variables}{A character vector containing valid variable names. Valid
variable names are stored in the HGSFile object and can be displayed with
<HGSFile>$variables, where <HGSFile> is the name of the HGSFile object.}

\item{blockNumber, solutionTime}{Specify one, but only one, of these two
variables to choose the time step from which data should be retrieved. Use
blockNumber to specify an index of the times steps that were output (e.g.,
to get data from the first time step that was output, specify '1', for the
second '2', etc.). Use solutionTime to specify the solution time from which
to retrieve data. (Note: Available solutionTimes are stored in the HGSFile
object and can be accessed using the \code{\link{HGSQueryBlocks}} function
with "SOLUTIONTIME" as a descriptor).}

\item{includeCoords}{Set to TRUE to include the X, Y, Z coordinates associated
with the selected variable.  As an alternative, X, Y, and/or Z can be
included the \code{variables} vector with other NODE centered variable
names; when \code{variables} containes names of CELLCENTERED variables, set
\code{includeCoords = T} to return the coordinates of cell centroids.}

\item{HGSdf}{S3 object of class "HGS.data.frame" created by calling
\code{\link{HGSGetData}}.}
}
\value{
\code{HGSGetData} returns an HGS.data.frame object of values
  associated with nodes (for NODECENTERED variables) or cells (for
  CELLCENTERED variables). There is one column for each variable requested,
  along with columns for the X, Y, and Z coordinates (in the spatial units
  determined when setting up the HGS model run) if includeCoords is TRUE.

  Use \code{attr(<HGS.data.frame_Object>, "varLoc")} to determine if
  variables in the HGS.data.frame are NODECENTERED and which are
  CELLCENTERED.

  \code{HGSCube} returns an "HGSCube" S3 object, which is 4-dimensional
  \code{array} where the first three dimensions are the node indices in the
  X, Y, and Z dimensions and the 4th dimension is the variable name. "NodeID"
  or "CellID" is a variable name added to the array automatically.  As with
  any \code{array} in R, the HGSCube can be sectioned using square brackets
  "[]". Note that "drop = F" can be a useful parameter after the dimensions
  specified within the square brackets to prevent collapsing with a single
  value of one dimension is specified.  See the help for square bracket by
  clicking this link: \code{\link{[}}

  The HGSCube object represents all output for a single model timestep.
  HGSCube inherits from HGSArray.  When HGSCube is sectioned with square
  brackets, it becomes simply an HGSArray object.
}
\description{
Get and manipulate output values associated with nodes or cells in an HGS
model run.
}
