% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{HGSGetData}
\alias{HGSGetData}
\title{Get data values from HGS files}
\usage{
HGSGetData(HGSFile, variables, blockNumber = NULL, solutionTime = NULL,
  X = integer(0), Y = integer(0), Z = integer(0),
  includeCoords = F)
}
\arguments{
\item{HGSFile}{An S3 object of class "HGSFile" created by calling
\code{\link{HGSFile}}.}

\item{variables}{A character vector containing valid variable names.  The
valid variable names are stored in the HGSFile object and cans be displayed
with <HGSFile>$variables, where <HGSFile> is the name of the HGSFile
object.}

\item{blockNumber, solutionTime}{Specify one of these two variables to return
data from a particular model time step that was output from HydroGeoSphere.
Use blockNumber to specify a serial number of the output block (e.g., 1, 2,
3, 4, etc.).  Use solutionTime to specify the specific solution time.
(Note: Available solutionTimes are stored in HGSFile objects and can be
accessed using the \code{\link{HGSQueryBlocks}} function with
"SOLUTIONTIME" as a descriptor).}

\item{X, Y, Z}{The index (i.e., the node number, not the spatial coordinates)
of nodes to be returned in each dimension.  Setting one or two of these
parameters involkes "slicing", returning either a vector or a matrix
representing an axis or a slice through the model.  See "Value" section,
below, for behavior depending on how many parameters are set.}

\item{includeCoords}{Set to T to include the X, Y, Z coordinates associated
with the selected variable.  Presently, this feature only works for
node-centered variables such as head and Depth2GWT.  A future release may
work with}
}
\value{
Either a data.frame, a vector, or a matrix, depending on the values
  for X, Y, and Z.

  HGSGetData will return a data.frame when no values are passed for X, Y, and
  Z (or X, Y, and Z are all equal to integer(0)).  The
  data.frame columns will contain the values for the variables requested,
  along with the x, y, and z coordinates (in spatial units used to run HGS)
  if includeCoords is TRUE.
}
\description{
Returns the data values associated with each node in an HGS model.
}
